+++ src/json_read_to_acmap.cpp
@@ -89,6 +89,8 @@ AcMap json_to_acmap(
     if(ag.HasMember("N")) map.antigens[i].set_name( ag["N"].GetString() );
     if(ag.HasMember("P")) map.antigens[i].set_passage( ag["P"].GetString() );
     if(ag.HasMember("c")) map.antigens[i].set_clade( parse<std::vector<std::string>>(ag["c"]) );
+    if(ag.HasMember("a")) map.antigens[i].set_annotations( parse<std::vector<std::string>>(ag["a"]) );
+    if(ag.HasMember("l")) map.antigens[i].set_labids( parse<std::vector<std::string>>(ag["l"]) );
     if(ag.HasMember("A")) map.antigens[i].set_sequence( ag["A"].GetString() );
     if(ag.HasMember("D")) map.antigens[i].set_date( ag["D"].GetString() );
     // set_reference
@@ -106,6 +108,7 @@ AcMap json_to_acmap(
     if(sr.HasMember("N")) map.sera[i].set_name( sr["N"].GetString() );
     if(sr.HasMember("P")) map.sera[i].set_passage( sr["P"].GetString() );
     if(sr.HasMember("c")) map.sera[i].set_clade( parse<std::vector<std::string>>(sr["c"]) );
+    if(sr.HasMember("a")) map.sera[i].set_annotations( parse<std::vector<std::string>>(sr["a"]) );
     if(sr.HasMember("A")) map.sera[i].set_sequence( sr["A"].GetString() );
     if(sr.HasMember("D")) map.sera[i].set_date( sr["D"].GetString() );
     if(sr.HasMember("I")) map.sera[i].set_id( sr["I"].GetString() );